perf(shader): mediump precision on fragment shaders — ~2x GPU throughput on mobile
Fragment shaders don't need highp for color math. Vertex shaders keep highp
for position precision. Recommended by ARM OpenGL ES best practices.

+++ apps/mobile/components/tower/BlockShader.ts
@@ -129,7 +129,7 @@ const vertexShader = /* glsl */ `
 `;
 
 const fragmentShader = /* glsl */ `
-  precision highp float;
+  precision mediump float;  // mediump sufficient for color math — ~2x faster on mobile GPU
 
   uniform float uTime;
   uniform vec3 uFogColor;
@@ -910,7 +910,7 @@ const glowVertexShader = /* glsl */ `
 `;
 
 const glowFragmentShader = /* glsl */ `
-  precision highp float;
+  precision mediump float;  // mediump sufficient for glow color math
 
   uniform float uTime;
